+++ packages/cervel/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @vercel/cervel
 
+## 0.1.26
+
+### Patch Changes
+
+- Updated dependencies [8dc4702]
+  - @vercel/backends@0.8.18
+
+## 0.1.25
+
+### Patch Changes
+
+- Updated dependencies [fdb6121]
+  - @vercel/backends@0.8.17
+
 ## 0.1.24
 
 ### Patch Changes