+++ packages/elysia/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @vercel/elysia
 
+## 0.1.97
+
+### Patch Changes
+
+- @vercel/node@5.8.21
+
+## 0.1.96
+
+### Patch Changes
+
+- @vercel/node@5.8.20
+
 ## 0.1.95
 
 ### Patch Changes